accessToken: Allow a user's clock to be wrong by up to 15 minutes
The default LiveKit access token signing sets the "not before" time to
the current date & time. If a user's machine does not have an accurate
clock, this can cause the token to fail.

Switch to signing the token ourselves using `jsonwebtoken` so we can
customize the parameters of the sign operation.

Set the "not before" time to 15 minutes before the current time to allow
for clocks that may be set inaccurately. An error may still happen if
the user's clock is off from the server clock by more than 15 minutes or
their timezone is set improperly.