devices: Capture errors when trying to get media devices
If devices can't be acquired, either due to missing permissions or the
use of an insecure connection, we don't want to just error out. Instead,
catch the error, log it, and return an empty list of devices.

Also switch to the LiveKit method for getting devices so browser
specific fixes made by them will be included.