simulcast: Adjust resolution defaults to fit FVTT's default sizing
Newer versions of FVTT have a default video resolution of 234x175.5.
This means that second simulcast layer was always the smallest resolution
being selected.

Increase the smallest simulcast resolution to fit these
new defaults so it is the only resolution that needs to be enabled
unless video windows are popped out for larger viewing. This also lets
us pick a video resolution a little bit higher for the second layer to
cover some more use cases.